refactor: move to exceptions
