+++ src/ape/cli/__init__.py
@@ -23,10 +23,9 @@ from ape.cli.options import (
     skip_confirmation_option,
 )
 from ape.cli.paramtype import AllFilePaths, Path
-from ape.cli.utils import Abort, abort
+from ape.cli.utils import Abort
 
 __all__ = [
-    "abort",
     "Abort",
     "account_option",
     "AccountAliasPromptChoice",
