+++ src/ape_console/plugin.py
@@ -8,8 +8,7 @@ from IPython.core.magic import Magics, line_magic, magics_class  # type: ignore
 
 import ape
 from ape._cli import cli
-from ape.cli import abort
-from ape.exceptions import ApeException, handle_ape_exception
+from ape.exceptions import ApeException, abort, handle_ape_exception
 from ape.logging import logger
 from ape.types import AddressType
 from ape.utils import cached_property
